pyproject.toml
[project]
name = "cfts"
description = "Paradigms for running auditory function tests using psiexperiment"
requires-python = ">=3.7"
license = {file = "LICENSE.txt"}
authors = [
    {name = "Brad Buran", email="buran@ohsu.edu"},
]
maintainers = [
    {name = "Brad Buran", email="buran@ohsu.edu"},
]
dependencies = [
    "cftsdata",
    "psiaudio",
    "psiexperiment",
    "enaml[qt5-pyqt] >=0.13.0",
    "enamlx >=0.6.2",
    "tqdm",
]
dynamic = ["version"]

[project.optional-dependencies]
ni = ["pydaqmx"]
tdt = ["tdtpy >=0.9.0"]
docs = ["sphinx", "sphinx_rtd_theme", "pygments-enaml"]
test = ["pytest", "pytest-console-scripts", "matplotlib"] # Matplotlib is used in some of the validation scripts.
zarr-backend = ["zarr"]
dev = ["coloredlogs"]

[project.scripts]
cfts = "cfts.app.main:cfts"

[build-system]
requires = ["setuptools>=61.2", "wheel", "setuptools_scm[toml]>=3.4.3"]
build-backend = "setuptools.build_meta"

[tool.setuptools_scm]
write_to = "cfts/version.py"
